requirements.txt
PyYAML>=3.13
paho-mqtt>=1.4.0